rrr: convert base/misc to module (rrr.misc); switch debugging/logging forward-decls to imports
misc.cpp becomes the rrr.misc module interface unit; misc.hpp deleted.
Includers updated:
  - src/rrr/rrr.hpp, src/rrr/base/all.hpp, src/rrr/base/threading.hpp:
    drop the include, add `import rrr.misc;`.
  - src/rrr/base/debugging.cpp, src/rrr/base/logging.cpp: drop their
    GMF forward-decls of get_exec_path / time_now_str; replace with
    `import rrr.misc;` in module purview now that misc is itself a
    module. Both modules' BMIs shrink as a result (rrr.debugging.pcm:
    25.7 → 24.6 MB).

Dropped from the public surface (unused outside this file):
  - `arraysize` macro and `ArraySizeHelper` template
  - `TIME_NOW_STR_SIZE` macro

Module BMI dedup visible: rrr.logging.pcm is 0.6 MB, rrr.misc.pcm is
11.3 MB; the others stay ~25 MB. Each module's BMI captures only its
own declarations plus references to imported modules.

Measurement (rrr + rpcbench, -j32, clean):
  +logging:   21.64 s, 102.5 MB PCM
  +misc:      30.14 s, 112.4 MB PCM, librrr.a 10.34 MB

+8.5 s wallclock jump in this step. Likely because misc is widely
fanned-out (imported transitively by debugging/logging/threading/all)
and now many TUs deserialize more module BMIs. Still 1.79x baseline,
well under the 3x abort threshold (~50 s).
