+++ src/rrr/CMakeLists.txt
@@ -20,6 +20,7 @@ set(RRR_MODULE_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/base/basetypes.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/base/debugging.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/base/logging.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/base/misc.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/base/strop.cpp
 )
 list(REMOVE_ITEM RRR_BENCH_SRC ${RRR_MODULE_SRC})
@@ -49,7 +50,7 @@ if(NOT TARGET rrr)
         ${CMAKE_CURRENT_SOURCE_DIR}/reactor/epoll_wrapper.cc
         ${CMAKE_CURRENT_SOURCE_DIR}/reactor/reactor.cc
         # base/logging.cpp omitted — module unit.
-        ${CMAKE_CURRENT_SOURCE_DIR}/base/misc.cpp
+        # base/misc.cpp omitted — module unit.
         # base/basetypes.cpp omitted — module unit (see RRR_MODULE_SRC).
         # base/debugging.cpp omitted — module unit.
         ${CMAKE_CURRENT_SOURCE_DIR}/base/threading.cpp
