+++ src/rrr/base/all.hpp
@@ -3,7 +3,6 @@
 //#include "unittest.hpp"
 
 
-#include "misc.hpp"
 #include "threading.hpp"
 
 // Imports go after textual #includes (libc++ rejects `import std; ...
@@ -11,4 +10,5 @@
 import rrr.basetypes;
 import rrr.debugging;
 import rrr.logging;
+import rrr.misc;
 import rrr.strop;
