+++ src/rrr/base/misc.cpp
@@ -1,45 +1,123 @@
+module;
 
-// import std; replacement — see <std_compat.hpp> for rationale.
-#include <stddef.h>
+#include <rusty/fn.hpp>
+#include <rusty/function.hpp>
 
-
-// @c-compat-added
-
-
-
-#include <stdio.h>
-#include <time.h>
 #include <limits.h>
-#include <unistd.h>
+#include <stdio.h>
 #include <stdlib.h>
 #include <sys/time.h>
+#include <time.h>
+#include <unistd.h>
 
-
-
-
-#include "misc.hpp"
-
-
-#include "../rrr.hpp"
+export module rrr.misc;
 
 import std;
+import rrr.basetypes;
 
-// External safety annotations for system functions used in this module
-// @external: {
-//   sysconf: [unsafe, (int) -> long]
-//   readlink: [unsafe, (const char*, char*, size_t) -> ssize_t]
-//   snprintf: [unsafe, (char*, size_t, const char*, ...) -> int]
-//   getpid: [unsafe, () -> pid_t]
-//   std::__atomic_base::load: [unsafe]
-//   std::__atomic_base::store: [unsafe]
-//   std::__atomic_base::fetch_add: [unsafe]
-//   std::__atomic_base::fetch_sub: [unsafe]
-// }
+export namespace rrr {
+
+inline uint64_t rdtsc() {
+#if defined(__i386__) || defined(__x86_64__)
+  uint32_t hi, lo;
+  __asm__ __volatile__("rdtsc" : "=a"(lo), "=d"(hi));
+  return (((uint64_t)hi) << 32) | ((uint64_t)lo);
+#elif defined(__aarch64__)
+  uint64_t val;
+  __asm__ __volatile__("mrs %0, cntvct_el0" : "=r"(val));
+  return val;
+#else
+  return 0;
+#endif
+}
+
+template <class T, class T1, class T2>
+inline T clamp(const T &v, const T1 &lower, const T2 &upper) {
+  if (v < lower) {
+    return lower;
+  }
+  if (v > upper) {
+    return upper;
+  }
+  return v;
+}
+
+// YYYY-MM-DD HH:MM:SS.mmm; caller-supplied buffer must be at least 24 bytes.
+void time_now_str(char *now);
+int get_ncpu();
+const char *get_exec_path();
+
+// NOTE: \n is stripped from input
+std::string getline(FILE *fp, char delim = '\n');
+
+template <class K, class V, class Map>
+inline void insert_into_map(Map &map, const K &key, const V &value) {
+  map.insert(typename Map::value_type(key, value));
+}
+
+template <class Container>
+typename std::reverse_iterator<typename Container::iterator>
+erase(Container &l,
+      typename std::reverse_iterator<typename Container::iterator> &rit) {
+  typename Container::iterator it = rit.base();
+  it--;
+  it = l.erase(it);
+  return std::reverse_iterator<typename Container::iterator>(it);
+}
+
+class Job {
+public:
+  virtual bool Ready() = 0;
+  virtual void Work() = 0;
+  virtual bool Done() = 0;
+  virtual ~Job() = default;
+};
+
+class OneTimeJob : public Job {
+ public:
+  OneTimeJob(rusty::Function<void()> func) : func_(std::move(func)) {
+  }
+  bool done_{false};
+  bool ready_{true};
+  rusty::Function<void()> func_{};
+  bool Ready() override { return ready_; }
+  bool Done() override { return done_; }
+  void Work() override {
+    ready_ = false;
+    func_();
+    done_ = true;
+  }
+};
+
+class FrequentJob : public Job {
+public:
+  uint64_t tm_last_ = 0;
+  uint64_t period_ = 0;
+
+  virtual ~FrequentJob() {}
+  virtual bool Ready() override {
+    uint64_t tm_now = rrr::Time::now();
+    uint64_t s = tm_now - tm_last_;
+    if (s > period_) {
+      tm_last_ = tm_now;
+      return true;
+    }
+    return false;
+  }
+
+  virtual bool Done() override {
+    return false;
+  }
+
+  virtual uint64_t get_last_time() { return tm_last_; }
+
+  virtual void set_period(uint64_t p) { period_ = p; }
+};
+
+} // export namespace rrr
 
 namespace rrr {
 
-// @unsafe - Direct pointer manipulation for performance
-// SAFETY: Caller must ensure str points to buffer with at least 'digits' bytes
 static void make_int(char* str, int val, int digits) {
     char* p = str + digits;
     for (int i = 0; i < digits; i++) {
@@ -50,12 +128,6 @@ static void make_int(char* str, int val, int digits) {
     }
 }
 
-// format time
-// inspired by the TPC-C benchmark from Evan Jones
-// strftime is slow because it ends up consulting timezone info
-// also, snprintf is slow
-// @unsafe - Direct buffer manipulation for performance
-// SAFETY: Caller must ensure 'now' points to buffer with at least TIME_NOW_STR_SIZE bytes
 void time_now_str(char* now) {
     time_t seconds_since_epoch = time(nullptr);
     struct tm local_calendar;
@@ -78,14 +150,10 @@ void time_now_str(char* now) {
     now[23] = '\0';
 }
 
-// @unsafe - Calls sysconf (external unsafe)
-// SAFETY: sysconf is POSIX-compliant, thread-safe
 int get_ncpu() {
     return sysconf(_SC_NPROCESSORS_ONLN);
 }
 
-// @unsafe - Calls readlink, snprintf, getpid (external unsafe)
-// SAFETY: Returns static buffer, thread-safe after first call
 const char* get_exec_path() {
     static char path[PATH_MAX];
     static bool ready = false;
@@ -103,9 +171,7 @@ const char* get_exec_path() {
     return path;
 }
 
-// @unsafe - Uses raw FILE* and manual memory management
-// SAFETY: FILE* must be valid; getdelim allocates buffer that we free
-std::string getline(FILE* fp, char delim /* =? */) {
+std::string getline(FILE* fp, char delim) {
     char* buf = nullptr;
     size_t n = 0;
     ssize_t n_read = ::getdelim(&buf, &n, delim, fp);
@@ -117,4 +183,4 @@ std::string getline(FILE* fp, char delim /* =? */) {
     return line;
 }
 
-} // namespace base
+} // namespace rrr
